Hackathon Interledger 2025/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
psycopg2-binary
python-dotenv
pydantic
twilio
phonenumbers
python-jose[cryptography]
bcrypt
pytest
pytest-asyncio
httpx
responses
ruff
